# optimize base conversions