--- a/base62.go
+++ b/base62.go
@@ -4,6 +4,8 @@ const (
 	// lexographic ordering (based on Unicode table) is 0-9A-Za-z
 	base62Characters = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
 	zeroString       = "000000000000000000000000000"
+	offsetUppercase  = 10
+	offsetLowercase  = 36
 )
 
 // Converts a base 62 byte into the number value that it represents.
@@ -12,9 +14,9 @@ func base62Value(digit byte) byte {
 	case digit >= '0' && digit <= '9':
 		return digit - '0'
 	case digit >= 'A' && digit <= 'Z':
-		return 10 + (digit - 'A')
+		return offsetUppercase + (digit - 'A')
 	default:
-		return 36 + (digit - 'a')
+		return offsetLowercase + (digit - 'a')
 	}
 }
 
@@ -33,6 +35,18 @@ func fastEncodeBase62(dst []byte, src []byte) {
 	// from because this is a O(N^2) algorithm, and we make N = N / 4 by working
 	// on 32 bits at a time.
 	parts := [5]uint32{
+		/*
+			These is an inlined version of:
+
+			  binary.BigEndian.Uint32(src[0:4]),
+			  binary.BigEndian.Uint32(src[4:8]),
+			  binary.BigEndian.Uint32(src[8:12]),
+			  binary.BigEndian.Uint32(src[12:16]),
+			  binary.BigEndian.Uint32(src[16:20]),
+
+			For some reason it gave better performance, may be caused by the
+			bound check that the Uint32 function does.
+		*/
 		uint32(src[0])<<24 | uint32(src[1])<<16 | uint32(src[2])<<8 | uint32(src[3]),
 		uint32(src[4])<<24 | uint32(src[5])<<16 | uint32(src[6])<<8 | uint32(src[7]),
 		uint32(src[8])<<24 | uint32(src[9])<<16 | uint32(src[10])<<8 | uint32(src[11]),