The PR adds optimized implementations of the base conversion routines. The trade off is on having an implementation that's a bit more complex, but performs much better.

Most of the efficiency is gained by working on 4-bytes words instead of single bytes when converting between base 256 and 62 (it made the routines 10 to 20x faster).
The algorithms also make assumptions about the size of the inputs and outputs, and inlining some operations which allowed for a couple extra savings.

Here were the benchmarks before the change:
```
BenchmarkAppend-4   	  300000	      4316 ns/op	       0 B/op	       0 allocs/op
BenchmarkString-4   	  300000	      4312 ns/op	      32 B/op	       1 allocs/op
BenchmarkParse-4    	  300000	      4745 ns/op	       0 B/op	       0 allocs/op
BenchmarkNew-4      	 1000000	      1237 ns/op	       0 B/op	       0 allocs/op
```

And after (I added a couple of them):
```
BenchmarkAppendEncodeBase62-4       	  300000	      4259 ns/op	       0 B/op	       0 allocs/op
BenchmarkAppendFastEncodeBase62-4   	 5000000	       329 ns/op	       0 B/op	       0 allocs/op
BenchmarkAppendDecodeBase62-4       	  300000	      4394 ns/op	       0 B/op	       0 allocs/op
BenchmarkAppendFastDecodeBase62-4   	10000000	       214 ns/op	       0 B/op	       0 allocs/op
BenchmarkAppend-4                   	 5000000	       333 ns/op	       0 B/op	       0 allocs/op
BenchmarkString-4                   	 5000000	       358 ns/op	      32 B/op	       1 allocs/op
BenchmarkParse-4                    	10000000	       229 ns/op	       0 B/op	       0 allocs/op
BenchmarkNew-4                      	 1000000	      1239 ns/op	       0 B/op	       0 allocs/op
```
There's probably room for more optimizations but considering the massive improvements in parsing and formatting KSUIDs, this seems like a good step forward already.

I kept the reference implementation in the test files for comparison purposes so we can always test the behavior of the optimized routiens against the original algorithms. 

Please take a look and let me know if anything should be changed.